checksums and gpg detached

+++ build.sh
@@ -13,44 +13,70 @@ source build.conf
 [ -x appimagetool-x86_64.AppImage ] || chmod +x ${appimgx}
 [ -f app/CSCSigner.jar ] || exit 1
 ls app/libs/*.jar >/dev/null || exit 1
+if [[ "$1" == "--sign" ]] ; then
+    [ -z "$(gpg -K)" ] && exit 1
+fi
 echo -e "\e[32mSanity checks passed\e[39m"
 
+# create output dir
+workdir=$(pwd)
+outputdir="${workdir}/out-${version}"
+[ -d ${outputdir} ] && rm -fr ${outputdir}
+mkdir ${outputdir} || exit 1
+
+# do the job for each type of cpu
 for arch in x86_64 i686 ; do
 
     [ "$arch" == "x86_64" ] && javatgz=jre-${jrever}-linux-x64.tar.gz 
     [ "$arch" == "i686" ] && javatgz=jre-${jrever}-linux-i586.tar.gz
+    appname="ValidDesk-${version}-${arch}"
 
     # create appdir
-    workdir=$(pwd)
-    appdir="${workdir}/ValidDesk-${version}-${arch}.AppDir"
+    appdir="${workdir}/${appname}.AppDir"
     [ -d ${appdir} ] && rm -fr ${appdir}
-    mkdir ${appdir}
+    mkdir ${appdir} || exit 1
 
     # copy template files
-    cd ${workdir}/Templates/
+    cd ${workdir}/Templates/ || exit 1
     cp AppRun ValidDesk ValidDesk.desktop ValidDesk.png ${appdir}
     echo -e "\e[32mCreated AppDir ${appdir}\e[39m"
 
     # extract jre
-    tar -C ${appdir} -xf ${workdir}/${javatgz}
-    cd ${appdir} 
+    tar -C ${appdir} -xf ${workdir}/${javatgz} || exit 1
+    cd ${appdir}
     jre="$(ls -1|grep -e '^jre[0-9.]*_[0-9]*$')"
-    mv $jre jre
+    mv $jre jre || exit 1
     echo -e "\e[32mExtracted JRE ${javatgz}\e[39m"
 
     # copy app
     cd ${workdir}/app
-    cp -r CSCSigner.jar libs ${appdir}
+    cp -r CSCSigner.jar libs ${appdir} || exit 1
     echo -e "\e[32mApp copied in the AppDir\e[39m"
 
     # build appimage
-    cd ${workdir}
+    cd ${workdir} || exit 1
+    output="${appname}.AppImage"
     appimgx="appimagetool-${arch}.AppImage"
-    ./${appimgx} "$@" -n ${appdir} ${output}
+    options="-n $1"
+    ./${appimgx} ${options} -n ${appdir} ${outputdir}/${output} || exit 1
     echo -e "\e[32mCreated appimage file ${output}\e[39m"
 
+    # checksums
+    cd ${outputdir} || exit 1
+    sha256sum ${output} > ${output}.sha256sum || exit 1
+    sha512sum ${output} > ${output}.sha512sum || exit 1
+
+    # signing
+    if [[ "$1" == "--sign" ]] ; then
+        gpg -s -b ${output} || exit 1
+        gpg -s -b -a ${output} || exit 1
+        echo -e "\e[32mCreated appimage file ${output}\e[39m"
+    fi
+
     # cleaning
-    rm -fr ${appdir}
+    cd ${workdir} || exit 1
+    rm -fr ${appdir} || exit 1
     echo -e "\e[32mAppDir removed.\e[39m"
+    echo -e "\e[32mSuccess!\e[39m"
 
 done
